js/xpconnect/tests/chrome/test_weakref.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=484459
-->
<window title="Weakrefs"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
    <iframe type="content" id="ifr">
    </iframe>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
    SimpleTest.waitForExplicitFinish();
    var util = window.windowUtils;
    var weakUtil = Cu.getWeakReference(util);
    util = null;

    function callback() {
      ok(weakUtil.get(), "Should still be alive here");
      SimpleTest.finish();
    }

    SpecialPowers.exactGC(callback);
  ]]></script>
</window>
